+++ src/Steffi.Renderers/Svg/Renderables/CircleRenderable.cs
@@ -3,10 +3,10 @@ using System.Xml.Linq;
 
 namespace Steffi.Renderers.Svg.Renderables;
 
-internal class CircleRenderable(int x, int y, int r,
+internal class CircleRenderable(decimal x, decimal y, decimal r,
 	IFillAndStroke fillAndStroke) : Renderable(x, y)
 {
-	public override (XElement Element, int Width, int Height) Render() => (
+	public override (XElement Element, decimal Width, decimal Height) Render() => (
 		Element: SvgBuilder.Circle(
 			cx: (X ?? 0) + r,
 			cy: (Y ?? 0) + r,
