+++ src/Steffi.Renderers/Svg/Renderables/PointsHelper.cs
@@ -4,10 +4,10 @@ namespace Steffi.Renderers.Svg.Renderables;
 
 internal static class PointsHelper
 {
-	internal static (int Width, int Height) MeasureBounds(List<Point2D> points)
+	internal static (decimal Width, decimal Height) MeasureBounds(List<Point2D> points)
 	{
-		int maxX = 0;
-		int maxY = 0;
+		decimal maxX = 0;
+		decimal maxY = 0;
 
 		foreach (var p in points)
 		{
@@ -18,7 +18,7 @@ internal static class PointsHelper
 		return (maxX, maxY);
 	}
 
-	internal static string ToSvgString(List<Point2D> points, int dx, int dy)
+	internal static string ToSvgString(List<Point2D> points, decimal dx, decimal dy)
 	{
 		if (dx == 0 && dy == 0)
 			return string.Join(' ', points);
